refactor: unify the htmlblock transformers
